APIs/twilio.com/twilio_flex_v1/1.55.0/python-aiohttp-client/tests/test_flex_v1_insights_questionnaires_category_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.flex_v1_insights_questionnaires_category import FlexV1InsightsQuestionnairesCategory
from openapi_server.models.list_insights_questionnaires_category_response import ListInsightsQuestionnairesCategoryResponse


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_create_insights_questionnaires_category(client):
    """Test case for create_insights_questionnaires_category

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'authorization': 'authorization_example',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'name': 'name_example'
        }
    response = await client.request(
        method='POST',
        path='/v1/Insights/QualityManagement/Categories',
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_insights_questionnaires_category(client):
    """Test case for delete_insights_questionnaires_category

    
    """
    headers = { 
        'authorization': 'authorization_example',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/v1/Insights/QualityManagement/Categories/{category_sid}'.format(category_sid='category_sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_insights_questionnaires_category(client):
    """Test case for list_insights_questionnaires_category

    
    """
    params = [('PageSize', 56),
                    ('Page', 56),
                    ('PageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'authorization': 'authorization_example',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/Insights/QualityManagement/Categories',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_update_insights_questionnaires_category(client):
    """Test case for update_insights_questionnaires_category

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'authorization': 'authorization_example',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'name': 'name_example'
        }
    response = await client.request(
        method='POST',
        path='/v1/Insights/QualityManagement/Categories/{category_sid}'.format(category_sid='category_sid_example'),
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

